requirements.txt
fastapi 
uvicorn 
python-dotenv
langchain
langchain-ollama
langchain-community
chromadb
unstructured
bs4
ollama
pandas